% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_wonder.R
\name{.load_cdc_wonder_raw}
\alias{.load_cdc_wonder_raw}
\title{Raw Data for Heart Disease Deaths by Month, 1999-2020, US Divisions}
\usage{
.load_cdc_wonder_raw()
}
\description{
Read the CDC WONDER data on monthly counts of deaths by Census division with
underlying cause of death determined to be related to heart disease according
to the following ICD-10 codes:
}
\details{
ICD-10 Codes:
  - I00-I02 (Acute rheumatic fever);
  - I05-I09 (Chronic rheumatic heart diseases);
  - I11.0 (Hypertensive heart disease with (congestive) heart failure);
  - I11.9 (Hypertensive heart disease without (congestive) heart failure);
  - I13.0 (Hypertensive heart and renal disease with (congestive) heart failure);
  - I13.1 (Hypertensive heart and renal disease with renal failure);
  - I13.2 (Hypertensive heart and renal disease with both (congestive) heart failure and renal failure);
  - I13.9 (Hypertensive heart and renal disease, unspecified);
  - I20-I25 (Ischaemic heart diseases);
  - I26-I28 (Pulmonary heart disease and diseases of pulmonary circulation);
  - I30-I51 (Other forms of heart disease)
}
